Implementing BFS Pathfinding
Using a Pure C# Class Node that will store the coordinates and the necessarry booleans to determine whether or not the current node is part of the pathfinding tree. This node is part of the GridManager Script which is dictionary in which the coordinates of those nodes will be stored.